backend/pyproject.toml
[tool.poetry]
name = "mitou2022-prototype"
version = "0.1.0"
description = ""
authors = ["add-bakkers <add.re.0828@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
Django = "3.2.12"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
